23/01/25

+++ WebStudy06_SpringSecurity/src/main/java/kr/or/ddit/member/service/MemberServiceImpl.java
@@ -7,10 +7,6 @@ import javax.annotation.Resource;
 import javax.inject.Inject;
 
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -24,13 +20,14 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 public class MemberServiceImpl implements MemberService {
-	// 결합력 최상
+	
 	@Inject
-	private MemberDAO memberDAO;
+	private MemberDAO memberDAO; // Layer와 의존관계 형성
+	
 	@Resource(name="authenticationManager")
-	private AuthenticationManager authenticationManager;
+	private AuthenticationManager authenticationManager; // 인증관련된 service
 	@Inject
-	private PasswordEncoder encoder;
+	PasswordEncoder encoder;
 	
 	@PostConstruct
 	public void init() {
@@ -40,15 +37,11 @@ public class MemberServiceImpl implements MemberService {
 	@Override
 	public ServiceResult createMember(MemberVO member) {
 		ServiceResult result = null;
-		try {
-			retrieveMember(member.getMemId());
-			result = ServiceResult.PKDUPLICATED;
-		}catch (UserNotFoundException e) {
-			String encoded = encoder.encode(member.getMemPass());
-			member.setMemPass(encoded);
-			int rowcnt = memberDAO.insertMember(member);
-			result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
-		}
+		String encoded = encoder.encode(member.getMemPass());
+		member.setMemPass(encoded);
+		int rowcnt = memberDAO.insertMember(member);
+		result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
+		
 		return result;
 	}
 
@@ -64,83 +57,38 @@ public class MemberServiceImpl implements MemberService {
 	}
 
 	@Override
-	public MemberVO retrieveMember(String memId) {
+	public MemberVO retrieveMember(String memId) throws UserNotFoundException {
 		MemberVO member = memberDAO.selectMember(memId);
-		if(member==null)
-			throw new UserNotFoundException(String.format(memId+"에 해당하는 사용자 없음."));
+		if(member == null) {
+			throw new UserNotFoundException(String.format(memId+"에 해당하는 사용자 없음"));
+		}
 		return member;
 	}
 
 	@Override
 	public ServiceResult modifyMember(MemberVO member) {
-//		MemberVO inputData = new MemberVO();
-//		inputData.setMemId(member.getMemId());
-//		inputData.setMemPass(member.getMemPass());
-//		
-//		ServiceResult result = authenticationManager.authenticate(inputData);
-//		if(ServiceResult.OK.equals(result)) {
-//			int rowcnt = memberDAO.updateMember(member);
-//			result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
-//		}
-//		return result;
-//		
-		ServiceResult result = null;
-		Authentication inputData = new UsernamePasswordAuthenticationToken(member.getMemId(), member.getMemPass());
-		try {
-			authenticationManager.authenticate(inputData);
+		MemberVO inputData = new MemberVO();
+		inputData.setMemId(member.getMemId());
+		inputData.setMemPass(member.getMemPass());
+		
+		ServiceResult result = authenticationManager.authenticate(inputData);
+		
+		if(ServiceResult.OK.equals(result)) {
 			int rowcnt = memberDAO.updateMember(member);
-			result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
-		}catch (UsernameNotFoundException e) {
-			result = ServiceResult.NOTEXIST;
-		}catch (AuthenticationException e) {
-			result = ServiceResult.INVALIDPASSWORD;
+			result =  rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
 		}
 		return result;
 	}
 
 	@Override
 	public ServiceResult removeMember(MemberVO member) {
-//		ServiceResult result = authService.authenticate(member);
-//		if(ServiceResult.OK.equals(result)) {
-//			int rowcnt = memberDAO.deleteMember(member.getMemId());
-//			result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
-//		}
-//		return result;
+		ServiceResult result = authenticationManager.authenticate(member);
 		
-		ServiceResult result = null;
-		Authentication inputData = 
-				new UsernamePasswordAuthenticationToken(member.getMemId(), member.getMemPass());
-		
-		try {
-			authenticationManager.authenticate(inputData);
+		if(ServiceResult.OK.equals(result)) {
 			int rowcnt = memberDAO.deleteMember(member.getMemId());
-			result = rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
-		}catch (UsernameNotFoundException e) {
-			result = ServiceResult.NOTEXIST;
-		}catch (AuthenticationException e) {
-			result = ServiceResult.INVALIDPASSWORD;
+			result =  rowcnt > 0 ? ServiceResult.OK : ServiceResult.FAIL;
 		}
 		return result;
 	}
 
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
